refactor: MyPage의 사이드바를 따로 구분해 상위 컴포넌트로 변경
