+++ src/App.tsx
@@ -1,23 +1,30 @@
 import React from "react"
-import "./App.css"
+import { RecoilRoot } from "recoil"
 import { ChakraProvider } from "@chakra-ui/react"
 import { BrowserRouter, Route, Routes } from "react-router-dom"
+
+// CSS
+import "./App.css"
+import theme from "@/styles/theme"
+
+// 기본 페이지
 import MainPage from "@/pages/MainPage/MainPage"
 import LoginPage from "@/pages/LoginPage/LoginPage"
 import Layout from "@/components/Layout/Layout"
 import RegisterPage from "@/pages/RegisterPage/RegisterPage"
-import MyProfile from "@/pages/ProfilePage/MyProfile"
-import MySchedule from "@/pages/MySchedulePage/MySchedule"
-import MyFavorite from "@/pages/MyFavoritePage/MyFavorite"
 import ScheduleSharePage from "@/pages/ScheduleSharePage/ScheduleSharePage"
-import theme from "@/styles/theme"
 import ScheduleDetailPage from "@/pages/ScheduleDetailPage/ScheduleDetailPage"
 import ReviewSharePage from "@/pages/ReviewSharePage/ReviewSharePage"
 import ReviewDetailPage from "@/pages/ReviewDetailPage/ReviewDetailPage"
-import SearchResultsPage from "./pages/SearchResultPage/SearchResultPage"
+import SearchResultsPage from "@/pages/SearchResultPage/SearchResultPage"
 import AddReviewPage from "@/pages/AddReviewPage/AddReviewPage"
 import CreateSchedulePage from "@/pages/CreateSchedulePage/CreateSchedulePage"
-import { RecoilRoot } from "recoil"
+
+// 마이페이지
+import MyPage from "@/pages/MyPage/MyPage"
+import MyProfile from "@/pages/MyPage/Profile/MyProfile"
+import MySchedule from "@/pages/MyPage/Schedule/MySchedule"
+import MyFavorite from "@/pages/MyPage/Favorite/MyFavorite"
 
 function App() {
   return (
@@ -29,9 +36,6 @@ function App() {
               <Route index element={<MainPage />} />
               <Route path="login" element={<LoginPage />} />
               <Route path="register" element={<RegisterPage />} />
-              <Route path="profile" element={<MyProfile />} />
-              <Route path="schedule" element={<MySchedule />} />
-              <Route path="favorite" element={<MyFavorite />} />
               <Route path="schedule_share" element={<ScheduleSharePage />} />
               <Route path="review_share" element={<ReviewSharePage />} />
               <Route path="schedule_share_detail" element={<ScheduleDetailPage />} />
@@ -39,6 +43,11 @@ function App() {
               <Route path="search" element={<SearchResultsPage />} />
               <Route path="addreview" element={<AddReviewPage initialStep={1} />} />
               <Route path="create_schedule" element={<CreateSchedulePage />} />
+              <Route path="mypage" element={<MyPage />}>
+                <Route path="profile" element={<MyProfile />} />
+                <Route path="schedule" element={<MySchedule />} />
+                <Route path="favorite" element={<MyFavorite />} />
+              </Route>
             </Route>
           </Routes>
         </BrowserRouter>
