+++ src/components/DropDownButton/DropDownButton.tsx
@@ -1,9 +1,10 @@
 import React, { useState } from "react"
-import { Box, Menu, MenuButton, MenuList } from "@chakra-ui/react"
+import { Box, BreadcrumbLink, Menu, MenuButton, MenuList } from "@chakra-ui/react"
 import { FiChevronDown, FiChevronUp } from "react-icons/fi"
 import { StyledMenuItem } from "@/components/DropDownButton/DropDownButton.style"
 import { MenuTitle } from "../NavBar/NavBar.style"
 import { DropDownButtonProps } from "./type"
+import { Link } from "react-router-dom"
 
 const DropDownButton: React.FC<DropDownButtonProps> = ({ title, contents }) => {
   const [isOpened, setIsOpened] = useState<boolean>(false)
@@ -25,7 +26,9 @@ const DropDownButton: React.FC<DropDownButtonProps> = ({ title, contents }) => {
 
         {contents.map((item, index) => (
           <Box key={index}>
-            <StyledMenuItem href="#">{item}</StyledMenuItem>
+            <Link to={`/${item[1]}`}>
+              <StyledMenuItem href="#">{item[0]}</StyledMenuItem>
+            </Link>
           </Box>
         ))}
       </MenuList>
