+++ src/components/DropDownButton/type.d.ts
@@ -1,4 +1,4 @@
 export interface DropDownButtonProps {
   title: string
-  contents: string[]
+  contents: string[][]
 }
