+++ src/pages/MyPage/MyPage.style.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const MyPageWrapper = styled.div`
+  width: 1280px;
+  display: flex;
+  justify-content: flex-start;
+`
