+++ src/styles/styles.ts
@@ -1,7 +1,1 @@
 import styled from "styled-components"
-
-export const MyPageWrapper = styled.div`
-  width: 1280px;
-  display: flex;
-  justify-content: center;
-`
